refactor(artifacts): Rename package for consistency
