+++ keel-artifact/src/test/kotlin/com/netflix/spinnaker/keel/artifacts/ArtifactDeployedListenerTests.kt
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.keel.artifact
+package com.netflix.spinnaker.keel.artifacts
 
 import com.netflix.spinnaker.keel.api.Resource
 import com.netflix.spinnaker.keel.api.id
