+++ keel-test/src/main/kotlin/com/netflix/spinnaker/keel/test/artifacts.kt
@@ -6,8 +6,8 @@ import com.netflix.spinnaker.keel.api.artifacts.DeliveryArtifact
 import com.netflix.spinnaker.keel.api.artifacts.VersioningStrategy
 import com.netflix.spinnaker.keel.api.plugins.ArtifactSupplier
 import com.netflix.spinnaker.keel.api.support.SpringEventPublisherBridge
-import com.netflix.spinnaker.keel.artifact.DebianArtifactSupplier
-import com.netflix.spinnaker.keel.artifact.DockerArtifactSupplier
+import com.netflix.spinnaker.keel.artifacts.DebianArtifactSupplier
+import com.netflix.spinnaker.keel.artifacts.DockerArtifactSupplier
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverService
 import io.mockk.mockk
